refactor: split papers, books, class materials
